+++ src/ralph/converters/xapi/constants.py
@@ -1,0 +1,29 @@
+"""Constants for xAPI specifications"""
+
+# Languages
+EN = "en"
+
+# xAPI activities
+XAPI_ACTIVITY_PAGE = "http://activitystrea.ms/schema/1.0/page"
+
+# xAPI verbs
+XAPI_VERB_VIEWED = "http://id.tincanapi.com/verb/viewed"
+
+# xAPI extensions
+XAPI_EXTENSION_ACCEPT_LANGUAGE = "https://www.fun-mooc.fr/extension/accept_language"
+XAPI_EXTENSION_AGENT = "https://www.fun-mooc.fr/extension/agent"
+XAPI_EXTENSION_COURSE_ID = "https://www.fun-mooc.fr/extension/course_id"
+XAPI_EXTENSION_COURSE_USER_TAGS = "https://www.fun-mooc.fr/extension/course_user_tags"
+XAPI_EXTENSION_HOST = "https://www.fun-mooc.fr/extension/host"
+XAPI_EXTENSION_IP = "https://www.fun-mooc.fr/extension/ip"
+XAPI_EXTENSION_ORG_ID = "https://www.fun-mooc.fr/extension/org_id"
+XAPI_EXTENSION_REFERER = "https://www.fun-mooc.fr/extension/referer"
+XAPI_EXTENSION_USER_ID = "https://www.fun-mooc.fr/extension/user_id"
+XAPI_EXTENSION_REQUEST = "https://www.fun-mooc.fr/extension/request"
+
+# Display Names
+VIEWED = "viewed"
+PAGE = "page"
+
+# Other
+HOME_PAGE = "https://www.fun-mooc.fr"